Move zttp-only tests to their own omitted module
